Print line breaks and enable links of comments in <description>
To make the comments as readable as on openstreetmap.org.

The "html" field of the JSON input is used now, which resembles the HTML
code of the note's content in openstreetmap.org.
It adds <p></p> for comment paragraphs, so a few <br> tags could be
removed. Links are detected and put in <a> tags.

Also avoid the additional "" quotes surrounding the comment, which would
look weird now for multiline comments.

Exemplary note:
https://www.openstreetmap.org/note/3613871
https://api.openstreetmap.org/api/0.6/notes.json?bbox=-0.0846723,51.5119196,-0.0846721,51.5119198&limit=1000&closed=-1